fix(sync): dial the invitation's admitters from the path that runs
#3772 added a dial of the invitation's `admitter_addrs` before a joiner
publishes, and #3782 made that dial budget per machine. Both landed in
`join_namespace.rs::await_namespace_ready`, which nothing calls: its only caller
is `join_and_wait_ready`, which has no callers, and neither does
`join_namespace_with_retry`. The live path is `join_group` ->
`request_namespace_join` -> `initiate_namespace_join`. Because the functions are
`pub` in a `pub mod`, dead-code analysis never said a word.

So the dialing half of the chain has never executed. The peer preference from
#3781 does run, but preference only helps among peers a stream can be opened to,
and `subscribed_peers` reports who is on the topic mesh — which is exactly what
has not converged in the case this path exists for.

The dial moves to `SyncManager::initiate_namespace_join`, before the join-stream
retry loop. That is where it belongs for two reasons beyond being reachable: the
admitter peer ids are already derived there for #3781's preference, from the same
`params.invitation_bytes`, and `SyncManager` holds `network_client` directly, so
`dial` is callable — the `SyncNetwork` trait has none, which is why this cannot
route through `sync_network`.

`group_admitter_routes` and its four tests move to
`sync/manager/namespace_join.rs`, beside the `open_namespace_join_stream` they
serve. Deliberately not left behind: #3795 proposes deleting the dead module, and
nothing of this should go with it.

One consequence worth noting. This call site is on the live path, so #3772's
follow-up — making `dial` report a closed mailbox instead of panicking — finally
matters: a dial racing shutdown now fails instead of taking the node down. That
hardening was arguably inert where it was.

Net 174 lines out of the dead module, 174 into the live one.
